requirements.txt
torch~=2.0.1
matplotlib~=3.5.1
torchvision~=0.15.2
shapely~=2.0.1
opencv-python~=4.8.0.74
numpy~=1.24.0
pandas~=1.3.5
scipy~=1.11.1
scikit-image~=0.20.0
geopandas~=0.13.2
pillow~=9.4.0
trimesh~=4.0.5